feat: implement minimal React agent using LangGraph defaults
- Replace custom React agent with LangGraph's create_react_agent
- Simplify planner with clearer task generation prompts
- Strip down tool_runner to minimal functionality
- Add instruction for React agent to ensure task completion means implementation
- Improve planner prompt to focus on deliverables, not setup steps
- Archive original implementations with underscore prefix

This significantly improves performance (2.5 min vs 7+ min) and reliability.
Tasks now complete with actual implementation rather than just planning.
